fix: add logo uri to lists with subnet info
