+++ actions/createSubnetGroupsList.js
@@ -1,31 +1,6 @@
 import fs from 'fs';
-import path from 'path';
-import { getChains } from './getChains.mjs';
-import { createChain } from './createChain.mjs';
-import {
-  SUBNETS_ROOT_PATH,
-  SUBNET_INFO_FILE,
-  SUBNET_GROUPS_FILE,
-  FORMATTED_SUBNET_GROUPS_FILE,
-  SUBNETS_FOLDER_URL,
-  SUBNET_LOGO_FILE,
-} from './constants.mjs';
-
-const allChains = getChains();
-
-function createSubnet(subnetId) {
-  const subnetInfo = JSON.parse(fs.readFileSync(path.resolve(SUBNETS_ROOT_PATH, subnetId, SUBNET_INFO_FILE), 'utf8'));
-  const chains = allChains
-    .filter((chainInfo) => chainInfo.subnetId === subnetId)
-    .map((chainInfo) => createChain(chainInfo.chainId));
-  const subnetLogoFilePath = `${SUBNETS_FOLDER_URL}/${subnetId}/${SUBNET_LOGO_FILE}`;
-
-  return {
-    ...subnetInfo,
-    chains,
-    logoUri: subnetLogoFilePath,
-  };
-}
+import { SUBNET_GROUPS_FILE, FORMATTED_SUBNET_GROUPS_FILE } from './constants.mjs';
+import { createSubnet } from './createSubnet.mjs';
 
 const subnetGroups = JSON.parse(fs.readFileSync(SUBNET_GROUPS_FILE, 'utf8'));
