+++ actions/createSubnetList.js
@@ -1,32 +1,11 @@
 import fs from 'fs';
-import path from 'path';
-import { getChains } from './getChains.mjs';
-import { createChain } from './createChain.mjs';
 import { getSubnets } from './getSubnets.mjs';
-import {
-  MAINNET_SUBNET_LIST_FILE,
-  SUBNETS_ROOT_PATH,
-  TESTNET_SUBNET_LIST_FILE,
-  SUBNET_LOGO_FILE,
-  SUBNETS_FOLDER_URL,
-} from './constants.mjs';
+import { MAINNET_SUBNET_LIST_FILE, TESTNET_SUBNET_LIST_FILE } from './constants.mjs';
+import { createSubnet } from './createSubnet.mjs';
 
 const allSubnets = getSubnets();
-const allChains = getChains();
 
-const formattedSubnets = Object.values(allSubnets).map((subnetInfo) => {
-  const hasPng = fs
-    .readdirSync(path.resolve(SUBNETS_ROOT_PATH, subnetInfo.subnetId), 'utf8')
-    .includes(SUBNET_LOGO_FILE);
-
-  return {
-    ...subnetInfo,
-    chains: allChains
-      .filter((chainInfo) => chainInfo.subnetId === subnetInfo.subnetId)
-      .map((chainInfo) => createChain(chainInfo.chainId)),
-    ...(hasPng && { logoUri: `${SUBNETS_FOLDER_URL}/${subnetInfo.subnetId}/${SUBNET_LOGO_FILE}` }),
-  };
-});
+const formattedSubnets = Object.values(allSubnets).map((subnetInfo) => createSubnet(subnetInfo.subnetId));
 
 const mainnetSubnets = formattedSubnets.filter(({ isTestnet }) => !isTestnet);
 const testnetSubnets = formattedSubnets.filter(({ isTestnet }) => isTestnet);
